+++ src/classmap.php
@@ -43,6 +43,7 @@ const CLASS_MAP = [
   'ABadCafe\\PDE\\Audio\\Sequence\\NoteOn' => '/audio/Pattern.php',
   'ABadCafe\\PDE\\Audio\\Sequence\\NoteOff' => '/audio/Pattern.php',
   'ABadCafe\\PDE\\Audio\\Sequence\\Pattern' => '/audio/Pattern.php',
+  'ABadCafe\\PDE\\Audio\\Player' => '/audio/Player.php',
   'ABadCafe\\PDE\\Audio\\IMachine' => '/audio/IMachine.php',
   'ABadCafe\\PDE\\Audio\\IPCMOutput' => '/audio/IPCMOutput.php',
   'ABadCafe\\PDE\\Audio\\IConfig' => '/audio/IConfig.php',
@@ -119,6 +120,8 @@ const CLASS_MAP = [
   'ABadCafe\\PDE\\Routine\\Base' => '/routine/common/Base.php',
   'ABadCafe\\PDE\\System\\ILoader' => '/system/ILoader.php',
   'ABadCafe\\PDE\\System\\IRateLimiter' => '/system/IRateLimiter.php',
+  'ABadCafe\\PDE\\System\\IAsynchronous' => '/system/IAsynchronous.php',
+  'ABadCafe\\PDE\\System\\TAsynchronous' => '/system/TAsynchronous.php',
   'ABadCafe\\PDE\\System\\Context' => '/system/Context.php',
   'ABadCafe\\PDE\\System\\RateLimiter\\Adaptive' => '/system/rate_limiter/Adaptive.php',
   'ABadCafe\\PDE\\System\\RateLimiter\\Simple' => '/system/rate_limiter/Simple.php',
